refactor: :art: login page display
