+++ src/components/Header.tsx
@@ -8,16 +8,18 @@ import Link from "next/link";
 
 export default function Header() {
   const pathname = usePathname();
+  if (pathname != "/signup") {
+    return <></>;
+  }
   return (
     <div>
       <Image
         src={BoschHeader}
         alt="hyperGraphic"
-        className="w-screen h-3 object-cover"
+        className="h-3 w-screen object-cover"
       />
       <span className="flex items-center justify-between pr-12">
-        <Image src={BoschBlackLogo} alt="Logo" className="dark:invert"/>
-        {pathname != "/signin" && <Link href={"/signin"}>Log in</Link>}
+        <Image src={BoschBlackLogo} alt="Logo" className="dark:invert" />
       </span>
     </div>
   );
